Update README.md
Hey there, I while in the middle of my due diligence before applying to the Solutions Engineer role, I noticed that new users need to use a search engine to find pnpm installation instructions. 
Not a huge change, but given the use and misuse of AI and sponsored results, users might end up in weird corners of the web.
Why not help them with a quick link to pnpm's installation page? :)